requirements.txt
azureml-core
pandas
mlflow